+++ CHANGELOG.md
@@ -6,6 +6,10 @@
   * Fixed issue with incorrect check for upgrade status of server
 * SPFarm
   * Removed SingleServer as ServerRole, since this is an invalid role.
+  * Handle case where null or empty CentralAdministrationUrl is passed in
+  * Move CentralAdministrationPort validation into parameter definition
+    to work with ReverseDsc
+  * Add NotNullOrEmpty parameter validation to CentralAdministrationUrl
   * Fixed error when changing developer dashboard display level.
 * SPFarmSolution
   * Fix for Web Application scoped solutions.
@@ -34,7 +38,7 @@
     the current unit test file.
 * SPFarm
   * Moved check for CentralAdministrationUrl is HTTP to Set method,
-    to prevent issues with ReverseDsc.
+    to prevent issues with ReverseDsc
 * SPInstall
   * Updated error code checks to force reboot.
 * SPProductUpdate